Get the binding sites for a Transcription Factor (TF)

Retrieve the binding sites and genome location for a given transcription factor.

get_binding_sites(regulondb, TF, seq.format = "table")

Arguments

regulondb
A regulondb object.

TF
name of the transcription factor.

seq.format
Default: table. Supported: fasta, wconsensus.

Value

A dataframe with the following columns corresponding to the TFBSs associated with the TF. If seq.format = "fasta", returns a string with the TFBS sequences with breakline character. If the transcription factor does not exist, returns NA.

ID (identifier)

left position

right position

strand

sequence

Examples

## Download the database if necessary
if(!file.exists(file.path(tempdir(), 'regulondb_sqlite3.db'))) {
    download_database(tempdir())
}

## Build the regulon db object
e_coli_regulondb <-
    regulondb(
        database_path = file.path(tempdir(), "regulondb_sqlite3.db"),
        organism = "E.coli",
        database_version = "1",
        genome_version = "1"
    )

## Get the binding sites for AraC
get_binding_sites(e_coli_regulondb, TF = "AraC")

#>                        ID    left   right  strand
#> 1  ECK120015742-araB-araC   70110   70126 forward
#> 2  ECK120012328-araB-araC   70131   70147 forward
#> 3  ECK120012320-araB-araC   70184   70200 reverse
#> 4  ECK120012323-araB-araC   70205   70221 reverse
#> 5  ECK120012603-araB-araC   70342   70358 reverse
#> 6       ECK120012335-araJ  412564  412580 reverse
#> 7       ECK120012992-araJ  412585  412601 reverse
#> 8       ECK120012990-araF 1986300 1986316 reverse
#> 9       ECK120012331-araF 1986321 1986337 reverse
#> 10      ECK120013555-araF 1986375 1986391 reverse
#> 11      ECK120012333-araF 1986396 1986412 reverse
#> 12      ECK120012915-araE 2982244 2982260 reverse
#> 13      ECK120012913-araE 2982265 2982281 reverse
#> 14      ECK125108641-xylA 3730824 3730840 reverse
#> 15      ECK125108643-xylA 3730847 3730863 reverse
#>                                 sequence
#> 1  ataaaaagcgTCAGGTAGGATCCGCTAatcttatgga
#> 2  ccgctaatctTATGGATAAAAATGCTAtggcatagca
#> 3  tctataatcaCGGCAGAAAAGTCCACAttgattattt
#> 4  caaaaacgcgTAACAAAAGTGTCTATAatcacggcag
#> 5  attcagagaaGAAACCAATTGTCCATAttgcatcaga
#> 6  tgaatacagaGGGGCGAATTATCTCTTggccttgctg
#> 7  aataactattCAGCAGGATAATGAATAcagaggggcg
#> 8  ccgctaaattTATGCACGTTCTCACTGtaattctgcg
#> 9  ggctttccctTATGTCTTTTCCCGCTAaatttatgca
#> 10 aggctaatctTATGGATTAATCTGCTGtgcattcgac
#> 11 ccaaagacaaCAAGGATTTCCAGGCTAatcttatgga
#> 12 tccatatttaTGCTGTTTCCGACCTGAcacctgcgtg
#> 13 cgacatgtcgCAGCAATTTAATCCATAtttatgctgt
#> 14 taacataattGAGCAACTGAAAGGGAGtgcccaatat
#> 15 attatctcaaTAGCAGTGTGAAATAACataattgagc

## Get the fasta binding sites sequences for AraC
fasta.seq <- get_binding_sites(e_coli_regulondb,
    TF = "AraC", seq.format = "fasta")
cat(fasta.seq)

#> >ECK120015742-araB-araC_70110:70126:forward
#> ataaaaagcgTCAGGTAGGATCCGCTAatcttatgga
#>  >ECK120012328-araB-araC_70131:70147:forward
#> ccgctaatctTATGGATAAAAATGCTAtggcatagca
#>  >ECK120012320-araB-araC_70184:70200:reverse
#> tctataatcaCGGCAGAAAAGTCCACAttgattattt
#>  >ECK120012323-araB-araC_70205:70221:reverse
#> caaaaacgcgTAACAAAAGTGTCTATAatcacggcag
#>  >ECK120012603-araB-araC_70342:70358:reverse
#> attcagagaaGAAACCAATTGTCCATAttgcatcaga
#>  >ECK120012335-araJ_412564:412580:reverse
#> tgaatacagaGGGGCGAATTATCTCTTggccttgctg
#>  >ECK120012992-araJ_412585:412601:reverse
#> aataactattCAGCAGGATAATGAATAcagaggggcg
#>  >ECK120012990-araF_1986300:1986316:reverse
#> ccgctaaattTATGCACGTTCTCACTGtaattctgcg
#>  >ECK120012331-araF_1986321:1986337:reverse
#> ggctttccctTATGTCTTTTCCCGCTAaatttatgca
#>  >ECK120013555-araF_1986375:1986391:reverse
#> aggctaatctTATGGATTAATCTGCTGtgcattcgac
#>  >ECK120012333-araF_1986396:1986412:reverse
#> ccaaagacaaCAAGGATTTCCAGGCTAatcttatgga
#>  >ECK120012915-araE_2982244:2982260:reverse
#> tccatatttaTGCTGTTTCCGACCTGAcacctgcgtg
#>  >ECK120012913-araE_2982265:2982281:reverse
#> cgacatgtcgCAGCAATTTAATCCATAtttatgctgt
#>  >ECK125108641-xylA_3730824:3730840:reverse
#> taacataattGAGCAACTGAAAGGGAGtgcccaatat
#>  >ECK125108643-xylA_3730847:3730863:reverse
#> attatctcaaTAGCAGTGTGAAATAACataattgagc