Plot genomic elements from regulonDB

Plot genomic elements from regulonDB

get_dna_objects(
  regulondb,
  genome = "eschColi_K12",
  from = 0,
  to = 4641628,
  elements = "gene"
)

Arguments

regulondb
A regulondb() object.

genome
A valid UCSC genome name.

from
A integer(1) specifying the left position.

to
A integer(1) specifying the right position.

elements
A character vector specifying which annotation elements to plot. It can be any from: "-10 promoter box", "-35 promoter box", "gene", "promoter", "Regulatory Interaction", "sRNA interaction", or "terminator".

Value

GenomicRanges::GRanges-class() object with the elements found.

Examples

## Connect to the RegulonDB database if necessary
if(!exists('regulondb_conn')) regulondb_conn <- connect_database()

#> snapshotDate(): 2020-02-12

#> regulondb_sqlite3_v10.6.2_DM.db already exists in local directory, setoverwrite = TRUE if you want to replace the existing file.

## Build the regulondb object
e_coli_regulondb <-
    regulondb(
        database_conn = regulondb_conn,
        organism = "chr",
        database_version = "1",
        genome_version = "1"
    )

## Get all genes from E. coli
get_dna_objects(e_coli_regulondb)

#> GRanges object with 4583 ranges and 4 metadata columns:
#>          seqnames          ranges strand |           id        type        name
#>             <Rle>       <IRanges>  <Rle> |  <character> <character> <character>
#>      [1]      chr 4265782-4266861      + | ECK120000001        gene         alr
#>      [2]      chr   795862-796551      + | ECK120000002        gene        modB
#>      [3]      chr 2531463-2532224      + | ECK120000003        gene        cysZ
#>      [4]      chr 3812731-3813951      + | ECK120000004        gene         dfp
#>      [5]      chr 4347404-4348744      - | ECK120000005        gene        dcuB
#>      ...      ...             ...    ... .          ...         ...         ...
#>   [4579]      chr 3032939-3033010      - | ECK125276531        gene        yqfH
#>   [4580]      chr   850332-850397      + | ECK125276532        gene        yliM
#>   [4581]      chr 1642122-1642211      + | ECK125276533        gene        ynfS
#>   [4582]      chr   568695-568844      + | ECK125276534        gene        ylcJ
#>   [4583]      chr   765050-765150      + | ECK125276535        gene        sdhX
#>                                                                                            description
#>                                                                                            <character>
#>      [1]                                                                            alanine racemase 1
#>      [2]                                                    molybdate ABC transporter membrane subunit
#>      [3]                                                               sulfate:H<sup>+</sup> symporter
#>      [4] fused 4'-phosphopantothenoylcysteine decarboxylase and phosphopantothenoylcysteine synthetase
#>      [5]                                                   anaerobic C4-dicarboxylate transporter DcuB
#>      ...                                                                                           ...
#>   [4579]                                                                                  protein YqfH
#>   [4580]                                                                                  protein YliM
#>   [4581]                                                                    Qin prophage; protein YnfS
#>   [4582]                                                                                  protein YlcJ
#>   [4583]                                                                     small regulatory RNA SdhX
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

## Get genes providing Genomic Ranges
get_dna_objects(e_coli_regulondb, from = 5000, to = 10000)

#> GRanges object with 6 ranges and 4 metadata columns:
#>       seqnames    ranges strand |           id        type        name
#>          <Rle> <IRanges>  <Rle> |  <character> <character> <character>
#>   [1]      chr 5683-6459      - | ECK120000009        gene        yaaA
#>   [2]      chr 3734-5020      + | ECK120000989        gene        thrC
#>   [3]      chr 9306-9893      + | ECK120001467        gene         mog
#>   [4]      chr 6529-7959      - | ECK120001508        gene        yaaJ
#>   [5]      chr 8238-9191      + | ECK120001509        gene        talB
#>   [6]      chr 5234-5530      + | ECK120002701        gene        yaaX
#>                                   description
#>                                   <character>
#>   [1] peroxide stress resistance protein YaaA
#>   [2]                      threonine synthase
#>   [3]       molybdopterin adenylyltransferase
#>   [4]               putative transporter YaaJ
#>   [5]                         transaldolase B
#>   [6]  DUF2502 domain-containing protein YaaX
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

## Get aditional elements within genomic positions
get_dna_objects(e_coli_regulondb, from = 5000, to = 10000, elements = c("gene", "promoter"))

#> GRanges object with 19 ranges and 4 metadata columns:
#>        seqnames    ranges strand |           id        type
#>           <Rle> <IRanges>  <Rle> |  <character> <character>
#>    [1]      chr 5683-6459      - | ECK120000009        gene
#>    [2]      chr 3734-5020      + | ECK120000989        gene
#>    [3]      chr 9306-9893      + | ECK120001467        gene
#>    [4]      chr 6529-7959      - | ECK120001508        gene
#>    [5]      chr 8238-9191      + | ECK120001509        gene
#>    ...      ...       ...    ... .          ...         ...
#>   [15]      chr      8229      + | ECK125230842    promoter
#>   [16]      chr      8232      + | ECK125230843    promoter
#>   [17]      chr      8352      + | ECK125230844    promoter
#>   [18]      chr      8638      + | ECK125230845    promoter
#>   [19]      chr      9277      + | ECK125230846    promoter
#>                                          name
#>                                   <character>
#>    [1]                                   yaaA
#>    [2]                                   thrC
#>    [3]                                    mog
#>    [4]                                   yaaJ
#>    [5]                                   talB
#>    ...                                    ...
#>   [15] TSS_21 (cluster) promoter with unknown
#>   [16]           TSS_22 promoter with unknown
#>   [17]           TSS_23 promoter with unknown
#>   [18]           TSS_24 promoter with unknown
#>   [19] TSS_25 (cluster) promoter with unknown
#>                                    description
#>                                    <character>
#>    [1] peroxide stress resistance protein YaaA
#>    [2]                      threonine synthase
#>    [3]       molybdopterin adenylyltransferase
#>    [4]               putative transporter YaaJ
#>    [5]                         transaldolase B
#>    ...                                     ...
#>   [15]  TSS_21 (cluster) promoter with unknown
#>   [16]            TSS_22 promoter with unknown
#>   [17]            TSS_23 promoter with unknown
#>   [18]            TSS_24 promoter with unknown
#>   [19]  TSS_25 (cluster) promoter with unknown
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths