Return complete regulatory network.

This function retrieves all the regulation networks in regulonDB between TF-TF, GENE-GENE or TF-GENE depending on the parameter 'type'.

get_regulatory_network(
  regulondb,
  regulator = NULL,
  type = "TF-GENE",
  cytograph = FALSE
)

Arguments

regulondb
A regulondb() object.

regulator
Name of TF or gene that acts as regulator. If NULL, the function retrieves all existent networks in the regulonDB.

type
"TF-GENE", "TF-TF", "GENE-GENE"

cytograph
If TRUE, displays network in Cytoscape. This option requires previous instalation and launch of Cytoscape.

Value

A regulondb_result object.

Author

Carmina Barberena Jonas, Jesús Emiliano Sotelo Fonseca, José Alquicira Hernández, Joselyn Chávez

Examples

## Connect to the RegulonDB database if necessary
if (!exists("regulondb_conn")) regulondb_conn <- connect_database()
#> snapshotDate(): 2021-08-03

## Build the regulon db object
e_coli_regulondb <-
    regulondb(
        database_conn = regulondb_conn,
        organism = "E.coli",
        database_version = "1",
        genome_version = "1"
    )

## Retrieve regulation of 'araC'
get_regulatory_network(e_coli_regulondb,
    regulator = "AraC",
    type = "TF-GENE"
)
#> regulondb_result with 14 rows and 3 columns
#>       regulator   regulated      effect
#>     <character> <character> <character>
#> 1          AraC        araA         +/-
#> 2          AraC        araB         +/-
#> 3          AraC        araC         +/-
#> 4          AraC        araD         +/-
#> 5          AraC        araE           +
#> ...         ...         ...         ...
#> 10         AraC        xylA           -
#> 11         AraC        xylB           -
#> 12         AraC        ygeA           +
#> 13         AraC        ydeM           -
#> 14         AraC        ydeN           -

## Retrieve all GENE-GENE networks
get_regulatory_network(e_coli_regulondb, type = "GENE-GENE")
#> regulondb_result with 4405 rows and 3 columns
#>                regulator   regulated      effect
#>              <character> <character> <character>
#> 1                   cpxR        acrD           +
#> 2                   cpxR        cpxR           +
#> 3                   cpxR        aroG           +
#> 4                   cpxR        cheA           -
#> 5                   cpxR        cheA           -
#> ...                  ...         ...         ...
#> 4401 bglJ,gadE,rcsA,rcsB        yqhG           +
#> 4402 bglJ,gadE,rcsA,rcsB        ygiZ           +
#> 4403 bglJ,gadE,rcsA,rcsB        rseD           +
#> 4404                  NA        accB           -
#> 4405                  NA        accC           -

## Retrieve TF-GENE network of AraC and display in Cytoscape
## Note that Cytospace needs to be open for this to work
cytoscape_present <- try(RCy3::cytoscapePing(), silent = TRUE)
#> Oh no! I can't find Cytoscape. RCy3 can not continue!
#> Please check that Cytoscape is running, CyREST is installed and your base.url parameter is correct.
if (!is(cytoscape_present, "try-error")) {
    get_regulatory_network(
        e_coli_regulondb,
        regulator = "AraC",
        type = "TF-GENE",
        cytograph = TRUE
    )
}