List attributes/fields from a dataset/table

List all attributes and their description of a dataset from RegulonDB. The result of this function may be used as parameter 'values' in list_attributes() function.

list_attributes(regulondb, dataset)

Arguments

regulondb

A regulondb() object.

dataset

Dataset of interest. The name should correspond to a table of the database.

Value

A character vector with the field names.

Author

Carmina Barberena Jonás, Jesús Emiliano Sotelo Fonseca, José Alquicira Hernández, Joselyn Chavez

Examples

## Connect to the RegulonDB database if necessary
if (!exists("regulondb_conn")) regulondb_conn <- connect_database()

## Build the regulon db object
e_coli_regulondb <-
    regulondb(
        database_conn = regulondb_conn,
        organism = "E.coli",
        database_version = "1",
        genome_version = "1"
    )

## List the transcription factor attributes
list_attributes(e_coli_regulondb, "TF")
#>  [1] "id"                          "name"                       
#>  [3] "synonyms"                    "encode_tf_gene_id"          
#>  [5] "encode_tf_operon_id"         "encode_tf_tu_id"            
#>  [7] "site_length"                 "symmetry"                   
#>  [9] "transcription_factor_family" "connectivity_class"         
#> [11] "sensing_class"               "global_function_all_tfbs"   
#> [13] "global_function_strong"      "note"                       
#> [15] "total_regulated_gene"        "total_effect_repressor_gene"
#> [17] "total_effect_activator_gene" "total_effect_dual_gene"     
#> [19] "total_effect_unknown_gene"   "total_tfbs"                 
#> [21] "total_ri"                    "total_regulated_tf"         
#> [23] "total_regulated_promoter"    "total_regulated_tu"         
#> [25] "total_regulated_operon"      "total_coregulated_sigma"    
#> [27] "total_active_conformation"   "total_coregulator"          
#> [29] "regulated_gene_info"         "regulated_tf_id"            
#> [31] "regulated_tf_name"           "regulated_tu_id"            
#> [33] "regulated_tu_name"           "regulated_operon_id"        
#> [35] "regulated_operon_name"       "regulated_pm_id"            
#> [37] "regulated_pm_name"           "coregulated_sigma"          
#> [39] "active_conformation_id"      "active_conformation_name"   
#> [41] "coregulator_id"              "coregulator_name"           
#> [43] "tfbs_unique"                 "regulation_info"            
#> [45] "evidence_reference"         

## List the operon attributes
list_attributes(e_coli_regulondb, "OPERON")
#>  [1] "id"                 "name"               "firstgeneposleft"  
#>  [4] "lastgeneposright"   "regulationposleft"  "regulationposright"
#>  [7] "strand"             "total_tu"           "total_gene"        
#> [10] "total_sigma"        "total_promoter"     "total_tfbs"        
#> [13] "total_ri"           "total_tf"           "total_terminator"  
#> [16] "tu_id"              "tu_name"            "gene_info"         
#> [19] "sigma_name"         "promoter_name"      "tf_name"           
#> [22] "ri_id"              "ri_info"            "tm_ids"