Matrix scan.

Scan sequences with one or several position-specific scoring matrices (PSSM) to identify instances of the corresponding motifs (putative sites). This function supports a variety of background models (Bernoulli, Markov chains of any order).

matrix_scan(
  sequence,
  matrix,
  sequence.format = "fasta",
  matrix.format = "tab",
  quick = FALSE,
  n.treatment = NULL,
  concensus.name = NULL,
  pseudo = 1,
  equi.pseudo = FALSE,
  top.matrices = NULL,
  background.model = NULL,
  background = NULL,
  background.input = NULL,
  background.window = NULL,
  markov = NULL,
  background.pseudo = NULL,
  return.fields = NULL,
  sort.distrib = FALSE,
  lth = NULL,
  uth = NULL,
  str = NULL,
  origin = NULL,
  decimals = NULL,
  crer.ids = NULL,
  organism = NULL
)

Arguments

sequence
Sequence(s) to scan - all the formats supported in RSAT can be used as input (default: fasta)

matrix
Matrix/ces to scan with. The matrix format is specified with the option matrix.format (see below) Default format: tab.

sequence.format
Supported fields:

fasta (default)

IG (Intelligenetics)

WC (wconsensus)

raw

matrix.format
Supported fields:

tab (default)

cb

consensus

gibbs

meme

assembly

quick
Delegates scanning to the C program matrix-scan-quick (developed by Matthieu Defrance). Evaluate if the quick mode is compatible with the selected output parameters, otherwise, run in the slower mode. Incompatible with - CRER scanning - window background model.

n.treatment
Treatment of N characters. These characters are often used in DNA sequences to represent undefined or masked nucleotides.

skip. N-containing regions are skipped.

score. N-containing regions are scored. The probability of an N is 1 for both the background model and the matrix. The N residues will thus contribute neither positively nor negatively to the weight score of the N-containing fragment. This option can be useful to detect sites which are at the border of N-containing regions, or in cases there are isolated N in the sequences.

pseudo
Pseudo-count for the matrix (default: 1). The pseudo-count reflects the possibility that residues that were not (yet) observed in the model might however be valid for future observations. The pseudo-count is used to compute the corrected residue frequencies.

equi.pseudo
If this option is called, the pseudo-weight is distributed in an equiprobable way between residues. By default, the pseudo-weight is distributed proportionally to residue priors, except for the -window option where equipseudo is default.

top.matrices
Only scan with the top # matrices per matrix file. This option is valid for some file formats containing multiple matrices where top matrices are generally more informative.

background.model
Background model is a tab-delimited specification of oligonucleotide frequencies.

background
To use a precalculated background model from RSAT. Works with organism and markov options. Type of sequences used as background model for estimating expected oligonucleotide frequencies. Supported:

upstream

upstream-noorf

upstream-noorf-rm

intergenic

background.input
Calculate background model from the input sequence set. This option requires to specify the order of the background model with the option markov.

background.window
Size of the sliding window for the background model calculation. This option requires to specify the order of the background model with the option markov (suitable for short order model only markov 0 or 1)

markov
Order of the markov chain for the background model. This option is incompatible with the option background.

background.pseudo
Pseudo frequency for the background models. Value must be a real between 0 and 1. If this option is not specified, the pseudo-frequency value depends on the background calculation. For background.input and background_window, the pseudo frequency is automatically calculated with the length (L) of the sequence following this formula: square-root of L divided by L+squareroot of L. For background.model, default value is 0.01. If the training sequence length (L) is known, the value can be set by background.pseudo option to square-root of L divided by L+squareroot of L.

return.fields
List of fields to return. Supported fields:

sites

rank

limits

normw

bg_model

matrix

freq_matrix

weight_matrix

distrib

sort.distrib
Sort score distribution by decreasing value of significance, if value TRUE. By default, the score distributions are sorted by score (weight).

lth
Lower threshold on some parameter. Format=list of 'parameter value'. Supported fields:

score

pval

eval

sig

normw

proba_M

proba_B

rank

crer_sites

crer_size

occ

occ_sum

inv_cum

exp_occ

occ_pval

occ_eval

occ_sig

occ_sig_rank

uth
Upper threshold on some parameter. Format=list of 'param value'. Supported parameters: same as lth.

str
Scan 1 or 2 strands for DNA sequences.

origin
Define the origin for the calculation of positions. origin -0 defines the end of each sequence as the origin. The matching positions are then negative values, providing the distance between the match and the end of the sequence.

decimals
Number of decimals displayed for the weight score.

crer.ids
Assign one separate feature ID per CRER. This option is convenient to distinguish separate CRERs.

organism
To use a precalculated background model from RSAT, choose the organism corresponding to the background model. Works with background and markov options.

consensus.name
Use the motif (degenerate) consensus as matrix name.

Value

an R object with results retrieved from RSAT

Examples

## Define the matrix
m <-
    '; MET4 matrix, from Gonze et al. (2005). Bioinformatics 21, 3490-500.
A |   7   9   0   0  16   0   1   0   0  11   6   9   6   1   8
C |   5   1   4  16   0  15   0   0   0   3   5   5   0   2   0
G |   4   4   1   0   0   0  15   0  16   0   3   0   0   2   0
T |   0   2  11   0   0   1   0  16   0   2   2   2  10  11   8
//
; MET31 matrix, from Gonze et al. (2005). Bioinformatics 21, 3490-500.
A |   3   6   9   6  14  18  16  18   2   0   0   0   1   3   8
C |   8   3   3   2   3   0   1   0  13   2   0   1   0   3   6
G |   4   3   4   8   0   0   1   0   2   0  17   1  17  11   1
T |   3   6   2   2   1   0   0   0   1  16   1  16   0   1   3
//
; PHO4 matrix, from Gonze et al. (2005). Bioinformatics 21, 3490-500.
A |   0   4   4   1   1  21   0   0   0   0   2   2   6   1   7
C |   2   7  12   6  20   0  20   0   1   0   5   5   8   4   6
G |   5   1   2  11   0   0   0  21   0  15   8   7   2  11   2
T |  14   9   3   3   0   0   1   0  20   6   6   7   5   5   6
'

## This is incomplete: