Generate random sequences

Generates random sequences with RSAT.

random_seq(
  n,
  seq.length,
  format = "fasta",
  type = NULL,
  seed = NULL,
  alphabet = NULL,
  expfreq = NULL,
  bg.model = NULL,
  organism = NULL,
  oligo.length = NULL,
  length.file = NULL,
  lw = NULL
)

Arguments

n
Number of sequences to generate

seq.length
Length of sequence to generate.

format
Format of sequence(s) to generate

type
Type of sequence(s) to generate (protein | DNA | other).

seed
Seed for the random generator.

alphabet
Alphabet. Must be followed by residue frequencies expressed precisely this way: a:t # c:g #

expfreq
Expected frequencies of oligomers in sequence(s) to generate. When this option is used, the sequences are generated according to a Markov chain.

bg.model
Background model. Automatically load a pre-calibrated expected frequency file from the RSAT genome distribution. When this option is used, the options organism and oligo.length are also required, to indicate the organism and the oligonucleotide length, respectively. This option is incompatible with the option expfreq. Type of sequences used as background model for estimating expected oligonucleotide frequencies (supported models):

equi (equiprobable residue frequencies; default).

upstream (all upstream sequences, allowing overlap with upstream ORFs. Requires to speciy a model organism).

upstream-noorf (all upstream sequences, preventing overlap with upstream ORFs. Requires to specify a model organism).

intergenic (intergenic frequencies. Whole set of intergenic regions, including upstream and downstream sequences. Requires to specify a model organism).

organism
Name of the organism when using a background model.

oligo.length
Length of oligomer when using a background model.

length.file
Length file. Allows to generate random sequences with the same lengths as a set of reference sequences.The length file contains two columns : sequence ID (ignored) and sequence length.

lw
Line width (0 for whole sequence on one line).

Value

a string with results retrieved from RSAT

Examples

## Obtain a random sequence from RSAT
res <- random_seq(
    n = 3,
    seq.length = 100,
    format = "fasta",
    type = "DNA",
    seed = 12,
    organism = "Escherichia_coli_K12"
)
cat(res)

#> rand_1	random sequence 1/3; type: dna; length: 100
#> GATCCCGTCACTGTACGCTTCCAGTGACGTATTGCCACCACCACCTACACTCCAAATGTACTCGTATTAG
#> AACCCGCTCCGTTCTCTTGCGCACAGTCCA
#>  rand_2	random sequence 2/3; type: dna; length: 100
#> GAATATATAAGTGGGGCGCTAATTATGGACTTTATTCCGGAACTCCTCGGGATTACTCTTTTTTGAACAC
#> GAATTCGGCGGCGGCCTCAGGCCGGTTTTG
#>  rand_3	random sequence 3/3; type: dna; length: 100
#> AGAAGGCTCCCATACCCTCCTGTGCTTCACGAACCGCTGTACGATTAAACATCTCTCGAAATCACGCTTA
#> GACATTCACAGCAAGCTACAGTTTGTAGCT