Connect to the regulondb database

If the file does not exist, this function downloads the RegulonDB SQLite database file into the path prior to loading it.

connect_database(
  path = tempdir(),
  overwrite = FALSE,
  ah = AnnotationHub::AnnotationHub()
)

Arguments

path
The directory where the database will be downloaded to.

overwrite
logical(1) specifying whether to overwrite the database file.

ah
An AnnotationtHub object AnnotationHub-class. Can be NULL if you want to force to use the backup download mechanism.

Value

An SQLiteConnection-class connection to the RegulonDB database.

Examples

## Connect to the RegulonDB database if necessary
if(!exists('regulondb_conn')) regulondb_conn <- connect_database()

#> snapshotDate(): 2020-03-31

#> regulondb_sqlite3_v10.6.2_DM.db already exists in local directory, setoverwrite = TRUE if you want to replace the existing file.

## Connect to the database without using AnnotationHub
regulondb_conn_noAH <- connect_database(ah = NULL, overwrite = TRUE)