Constructs a logical condition to query database

Given a list of filters, this function builds a logical condition to query database using intervals. The output is used in existing_intervals() and non_existing_intervals().

existing_partial_match(filters, partialmatch, operator)

Arguments

filters
List of filters to be used. The names should correspond to the attribute and the values correspond to the condition for selection.

partialmatch
name of the condition(s) with a string pattern for full or partial match in the query.

operator
A string indicating if all the filters (AND) or some of them (OR) should be met.

Value

A character(1) with the sql logical condition to query the dataset.

Examples

## Build the SQL query for existing partial matches for ara
existing_partial_match(
    filters = list(
        name = c("ara"),
        strand = c("forward"),
        posright = c("2000", "40000")
    ),
    partialmatch = "name",
    operator = "AND"
)

#> [1] "(name like  '%ara%')"